PR (#18)
* chore(selfbot.core.telegram): set name for dispatching task

* refactor(selfbot.utils): replace `get_event_loop` with `get_running_loop`

* python 3.12 deprecates get_event_loop in coroutine context.
* this change replaces it with get_running_loop for better compatibility and clarity.

* feat(selfbot.utils): `fmtstr`

* output message formatter

* adeed `rflag` for restart indication

* bad logic

* fix(selfbot.core.telegram): `restart`

---------

+++ selfbot/modules/system.py
@@ -75,6 +75,13 @@ class System(Module):
             with open(file, "w") as f:
                 f.write(text)
 
+        if getattr(self.client, "restart", None):
+            return await event.edit_message_text(
+                "<code>Restart is Called</code>", reply_markup=ikm(("Close", b"0"))
+            )
+
+        setattr(self.client, "restart", True)
+
         if os.path.isdir(".git"):
             await shell(
                 "git fetch ; git reset --hard origin/{}".format(
@@ -97,6 +104,6 @@ class System(Module):
         await event.edit_message_text("<code>Restarting...</code>")
 
         try:
-            self.client.__idle__.cancel()
+            self.client.__idle__.set()
         finally:
             os.execv(sys.executable, (sys.executable, "-m", "selfbot"))
